+++ start-local-node.sh
@@ -55,4 +55,4 @@ chain-maind genesis collect-gentxs
 chain-maind genesis validate
 
 # Start the node (remove the --pruning=nothing flag if historical queries are not needed)
-chain-maind start --pruning=nothing --rpc.unsafe --trace --minimum-gas-prices 1000basecro --api.enable=true --api.address="tcp://0.0.0.0:1317"
+chain-maind start --pruning=nothing --rpc.unsafe --trace --minimum-gas-prices 1000basecro --api.enable=true --api.swagger=true --api.address="tcp://0.0.0.0:1317"